refactor(icons): split glyph registry from the sized ServiceIcon primitive
Move the app-facing, size-tokened `<ServiceIcon>` out of the design-owned
glyph registry into its own engineering-owned module, so the boundary
matches ownership and survives the icon planning-worktree sync.

- `components/service-icons/` is now a PURE glyph registry (design-owned,
  synced): `SERVICE_ICONS`, `SPEC_CATALOG_SLUGS`, `FallbackIcon`. No app
  tokens, no `cn`, no sizes.
- `components/service-icon.tsx` (new, engineering-owned) is the primitive:
  `<ServiceIcon slug=… size="xs|sm|md|lg" />` over the registry. All 14
  call sites import from here.
- Dropped the speculative bare-name prefix guessing; require the catalog
  slug the call sites already pass.

Also: the primitive now owns its box unconditionally (`!`-important
sizes + composite `[&>svg]:!h-full`). shadcn containers like
`DropdownMenuItem` ship a blanket `[&_svg]:h-3.5 w-3.5` that was shrinking
service icons to 14px and mangling composite badge layout in the SA
connect-service dropdown; important sizing defeats that everywhere.

Regenerated the CLI wizard bundle.
