+++ cli/src/wizard/bundle-meta/index.hash
@@ -1,1 +1,1 @@
-4edcebbdf0a45b8365be6098b7138b949dcc3aa0d6d5b7942959a1b2d37ba473
+d1148ab33b53c7b4006f286bb01e9f22ff8f614fc61074a2e08548b5d155df10
