+++ frontend/src/pages/service-detail.tsx
@@ -36,7 +36,7 @@ import { CopyableField } from "@/components/shared/copyable-field";
 import { DefaultHeadersEditor } from "@/components/shared/default-headers-editor";
 import { OidcCredentialsSection } from "@/components/dashboard/oidc-credentials-section";
 import { EndpointList } from "@/components/dashboard/endpoint-list";
-import { ServiceIcon } from "@/components/service-icons";
+import { ServiceIcon } from "@/components/service-icon";
 import { McpConnectionInfo } from "@/components/dashboard/mcp-connection-info";
 import { SshServiceInstructions } from "@/components/dashboard/ssh-service-instructions";
 import { ServiceRequirementsView } from "@/components/dashboard/service-requirements-editor";
